OCM-7184 | feat: Moving create machinepool and nodepool functions from cmd -> pkg
