--- cmd/create/machinepool/machinepool_suite_test.go
@@ -1,13 +1,0 @@
-package machinepool
-
-import (
-	"testing"
-
-	. "github.com/onsi/ginkgo/v2"
-	. "github.com/onsi/gomega"
-)
-
-func TestIdp(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "MachinePool Suite")
-}
