+++ pkg/machinepool/machinepool_mock.go
@@ -5,7 +5,6 @@
 //
 //	mockgen -source=machinepool.go -package=machinepool -destination=machinepool_mock.go
 //
-
 // Package machinepool is a generated GoMock package.
 package machinepool
 
@@ -14,6 +13,7 @@ import (
 
 	v1 "github.com/openshift-online/ocm-sdk-go/clustersmgmt/v1"
 	rosa "github.com/openshift/rosa/pkg/rosa"
+	cobra "github.com/spf13/cobra"
 	gomock "go.uber.org/mock/gomock"
 )
 
@@ -40,16 +40,72 @@ func (m *MockMachinePoolService) EXPECT() *MockMachinePoolServiceMockRecorder {
 	return m.recorder
 }
 
-// DescribeMachinePool mocks base method.
-func (m *MockMachinePoolService) DescribeMachinePool(r *rosa.Runtime, cluster *v1.Cluster, clusterKey string, isHypershift bool, machinePoolId string) error {
+// AddMachinePool mocks base method.
+func (m *MockMachinePoolService) AddMachinePool(cmd *cobra.Command, clusterKey string, cluster *v1.Cluster, r *rosa.Runtime, args *MachinePoolArgs) error {
 	m.ctrl.T.Helper()
-	ret := m.ctrl.Call(m, "DescribeMachinePool", r, cluster, clusterKey, isHypershift, machinePoolId)
+	ret := m.ctrl.Call(m, "AddMachinePool", cmd, clusterKey, cluster, r, args)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// AddMachinePool indicates an expected call of AddMachinePool.
+func (mr *MockMachinePoolServiceMockRecorder) AddMachinePool(cmd, clusterKey, cluster, r, args any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddMachinePool", reflect.TypeOf((*MockMachinePoolService)(nil).AddMachinePool), cmd, clusterKey, cluster, r, args)
+}
+
+// AddNodePool mocks base method.
+func (m *MockMachinePoolService) AddNodePool(cmd *cobra.Command, clusterKey string, cluster *v1.Cluster, r *rosa.Runtime, args *MachinePoolArgs) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "AddNodePool", cmd, clusterKey, cluster, r, args)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// AddNodePool indicates an expected call of AddNodePool.
+func (mr *MockMachinePoolServiceMockRecorder) AddNodePool(cmd, clusterKey, cluster, r, args any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "AddNodePool", reflect.TypeOf((*MockMachinePoolService)(nil).AddNodePool), cmd, clusterKey, cluster, r, args)
+}
+
+// DeleteMachinePool mocks base method.
+func (m *MockMachinePoolService) DeleteMachinePool(r *rosa.Runtime, machinePoolId, clusterKey string, cluster *v1.Cluster) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DeleteMachinePool", r, machinePoolId, clusterKey, cluster)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// DeleteMachinePool indicates an expected call of DeleteMachinePool.
+func (mr *MockMachinePoolServiceMockRecorder) DeleteMachinePool(r, machinePoolId, clusterKey, cluster any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DeleteMachinePool", reflect.TypeOf((*MockMachinePoolService)(nil).DeleteMachinePool), r, machinePoolId, clusterKey, cluster)
+}
+
+// DescribeMachinePool mocks base method.
+func (m *MockMachinePoolService) DescribeMachinePool(r *rosa.Runtime, cluster *v1.Cluster, clusterKey, machinePoolId string) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "DescribeMachinePool", r, cluster, clusterKey, machinePoolId)
 	ret0, _ := ret[0].(error)
 	return ret0
 }
 
 // DescribeMachinePool indicates an expected call of DescribeMachinePool.
-func (mr *MockMachinePoolServiceMockRecorder) DescribeMachinePool(r, cluster, clusterKey, isHypershift, machinePoolId any) *gomock.Call {
+func (mr *MockMachinePoolServiceMockRecorder) DescribeMachinePool(r, cluster, clusterKey, machinePoolId any) *gomock.Call {
 	mr.mock.ctrl.T.Helper()
-	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DescribeMachinePool", reflect.TypeOf((*MockMachinePoolService)(nil).DescribeMachinePool), r, cluster, clusterKey, isHypershift, machinePoolId)
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "DescribeMachinePool", reflect.TypeOf((*MockMachinePoolService)(nil).DescribeMachinePool), r, cluster, clusterKey, machinePoolId)
+}
+
+// ListMachinePools mocks base method.
+func (m *MockMachinePoolService) ListMachinePools(r *rosa.Runtime, clusterKey string, cluster *v1.Cluster) error {
+	m.ctrl.T.Helper()
+	ret := m.ctrl.Call(m, "ListMachinePools", r, clusterKey, cluster)
+	ret0, _ := ret[0].(error)
+	return ret0
+}
+
+// ListMachinePools indicates an expected call of ListMachinePools.
+func (mr *MockMachinePoolServiceMockRecorder) ListMachinePools(r, clusterKey, cluster any) *gomock.Call {
+	mr.mock.ctrl.T.Helper()
+	return mr.mock.ctrl.RecordCallWithMethodType(mr.mock, "ListMachinePools", reflect.TypeOf((*MockMachinePoolService)(nil).ListMachinePools), r, clusterKey, cluster)
 }
